+++ packages/ext-react-classic-renderwhenready/src/index.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import App1 from './App1'
+import App2 from './App2'
+import { render } from '@sencha/ext-react'
+import { ExtReact } from '@sencha/ext-react'
+
+render(<ExtReact><App1/></ExtReact>, document.getElementById('div1'))
+render(<App2/>, document.getElementById('div2'))