+++ packages/ext-react-demo/src/index.js
@@ -8,7 +8,13 @@ import { ExtReact } from '@sencha/ext-react'
 //import App from './D3/AppD3'
 //import App from './Progress/AppProgress'
 import App from './Button/AppButton'
+//import App from './Router/App'
 
+Ext.require('Ext.panel.Collapser')
+
+var store = Ext.create('Ext.data.Store', {
+  data: {a:'1'}
+})
 
 let viewport