+++ packages/ext-react-froala-example/webpack.config.js
@@ -1,25 +1,32 @@
 const webpack = require('webpack');
 const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin')
-const ExtReactWebpackPlugin = require('@sencha/ext-react-webpack-plugin')
+const ExtWebpackPlugin = require('@sencha/ext-react-webpack-plugin')
 const portfinder = require('portfinder')
 const sourcePath = path.join(__dirname, './src');
 
 module.exports = function (env) {
-  portfinder.basePort = (env && env.port) || 1962; // the default port to use
+  var buildprofile = env.profile || process.env.npm_package_extbuild_defaultprofile
+  var buildenvironment = env.environment || process.env.npm_package_extbuild_defaultenvironment
+  var buildverbose = env.verbose || process.env.npm_package_extbuild_defaultverbose
+  if (buildprofile == 'all') { buildprofile = '' }
+  const isProd = buildenvironment === 'production'
+
+  portfinder.basePort = (env && env.port) || 1962
   return portfinder.getPortPromise().then(port => {
-    const nodeEnv = env && env.prod ? 'production' : 'development';
-    const isProd = nodeEnv === 'production'
-    const local = env && env.local
     const plugins = [
       new HtmlWebpackPlugin({
         template: 'index.html',
         hash: true
       }), 
-      new ExtReactWebpackPlugin({
+      new ExtWebpackPlugin({
+        framework: 'react',
         port: port,
-        production: isProd,
-        treeShaking: false
+        profile: buildprofile, 
+        environment: buildenvironment, 
+        verbose: buildverbose,
+        theme: 'theme-material',
+        packages: []
       }),
 
       // 1. Froala 
@@ -42,8 +49,6 @@ module.exports = function (env) {
       devtool: isProd ? 'source-map' : 'cheap-module-source-map',
       context: sourcePath,
       entry: {
-        'vendor': ['react', 'prop-types', 'react-dom', 'react-router-dom', 'history'],
-        'ext-react': ['@sencha/ext-react'],
         'app': ['./index.js']
       },
       output: {